requirements/base.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosmtplib==1.1.6
aniso8601==9.0.1
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.1
astroid==2.11.4
asttokens==2.0.5
async-timeout==4.0.2
Authlib==1.0.1
backcall==0.2.0
bcrypt==3.2.2
blinker==1.4
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
decorator==5.1.1
dill==0.3.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
executing==0.8.3
fakeredis==1.7.1
fastapi==0.76.0
fastapi-mail==0.4.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
httpx-oauth==0.6.0
hyperframe==6.0.1
idna==3.3
ipython==8.3.0
isort==5.10.1
itsdangerous==2.1.2
jdatetime==4.1.0
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
limits==1.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
motor==3.0.0
msgpack==1.0.3
oauthlib==3.2.0
orjson==3.6.8
packaging==21.3
parso==0.8.3
passlib==1.7.4
persian-tools==0.0.10
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
pylint==2.13.8
pymongo==4.1.1
pyparsing==3.0.8
python-dotenv==0.20.0
python-jose==3.3.0
python-magic==0.4.25
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.2.0
starlette==0.18.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
xlrd==2.0.1
iteration-utilities==0.11.0
